+++ database/migrations/2021_02_10_000000_version_531.php
@@ -1,0 +1,30 @@
+<?php
+
+use BT\Modules\Settings\Models\Setting;
+use Illuminate\Database\Migrations\Migration;
+
+class Version531 extends Migration
+{
+
+    /**
+     * Run the migrations.
+     * @table payments_custom
+     *
+     * @return void
+     */
+    public function up()
+    {
+        Setting::saveByKey('schedulerFcTodaybgColor', '#FFF9DE');
+        Setting::saveByKey('version', '5.3.1');
+    }
+
+    /**
+     * Reverse the migrations.
+     *
+     * @return void
+     */
+     public function down()
+     {
+       //
+     }
+}